prepare monitor for deployment
