--- cmd/mirroring/index.html
@@ -1,27 +1,0 @@
-<!DOCTYPE html>
- <html lang="en">
- <head>
-     <meta charset="UTF-8">
-     <meta http-equiv="X-UA-Compatible" content="IE=edge">
-     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-     <script src="wasm_exec.js"></script>
-     <script>
-         if (!WebAssembly.instantiateStreaming) {
-             WebAssembly.instantiateStreaming = async (resp, importObject) => {
-                 const source = await (await resp).arrayBuffer();
-                 return await WebAssembly.instantiate(source, importObject);
-             };
-         }
-
-         const go = new Go();
-         WebAssembly.instantiateStreaming(fetch("monitor.wasm"), go.importObject).then((result) => {
-             go.run(result.instance);
-         });
-     </script>
-
-     <title>Document</title>
- </head>
- <body>
-
- </body>
- </html>
