+++ go.mod
@@ -4,10 +4,11 @@ go 1.16
 
 require (
 	github.com/go-openapi/runtime v0.19.30
-	github.com/go-sql-driver/mysql v1.6.0 // indirect
+	github.com/go-sql-driver/mysql v1.6.0
 	github.com/google/trillian v1.3.14-0.20210713114448-df474653733c
-	github.com/mattn/go-sqlite3 v1.14.11 // indirect
+	github.com/mattn/go-sqlite3 v1.14.11
 	github.com/sigstore/rekor v0.3.0
 	github.com/sigstore/sigstore v0.0.0-20210713222344-1fee53516622
 	github.com/spf13/viper v1.8.1
+	go.mongodb.org/mongo-driver v1.9.0 // indirect
 )
